Build grid and list DataView baseline (#103)
